requirements.txt
flask==2.3.3
python-dotenv==1.0.1
langchain==0.1.16
google-generativeai==0.3.2
faiss-cpu==1.7.4
gunicorn==21.2.0
